bug fixes & removed designslice.js
