+++ client/src/utils/reducers/appSlice.js
@@ -2,9 +2,7 @@ import { createSlice } from '@reduxjs/toolkit';
 
 const initialState = {
   message: null,
-  activeStep: 0,
   page: 'HOME',
-  userDesigns: [],
 };
 
 const appSlice = createSlice({
@@ -17,38 +15,14 @@ const appSlice = createSlice({
     resetMessage: (state) => {
       state.message = null;
     },
-    nextStep: (state) => {
-      state.activeStep += 1;
-    },
-    prevStep: (state) => {
-      state.activeStep -= 1;
-    },
-    resetStep: (state) => {
-      state.activeStep = 0;
-    },
     goToPage: (state, action) => {
       state.page = action.payload;
-    },
-    setUserDesigns: (state, action) => {
-      state.userDesigns = action.payload;
-    },
-    addToUserDesign: (state, action) => {
-      state.userDesigns = [...state.userDesigns, action.payload];
     },
     resetApp: (state) => initialState,
   },
 });
 
-export const {
-  setMessage,
-  resetMessage,
-  nextStep,
-  prevStep,
-  resetStep,
-  goToPage,
-  setUserDesigns,
-  addToUserDesign,
-  resetApp,
-} = appSlice.actions;
+export const { setMessage, resetMessage, goToPage, resetApp } =
+  appSlice.actions;
 
 export default appSlice.reducer;
